refactor selected file title data propagation
