+++ frontend/src/atoms/index.ts
@@ -6,8 +6,6 @@ import { DbViewEnum, DEFAULT_DB_VIEW } from "@utils/constants";
 import { atom } from "jotai";
 import { atomWithStorage } from "jotai/utils";
 
-import { CurrentDbFileAtom } from "./types";
-
 /**
  * GENERAL
  */
@@ -16,8 +14,6 @@ export const currentDbViewAtom = atomWithStorage<DbViewEnum>(
   "db-view",
   DEFAULT_DB_VIEW,
 );
-
-export const currentDbFileAtom = atom<CurrentDbFileAtom>(null);
 
 /**
  * SEARCH
