+++ frontend/src/features/numbersView/NumbersTable.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { TableComponents, TableVirtuoso } from "react-virtuoso";
-import { currentDbFileAtom } from "@atoms";
 import { Box, Skeleton, TableCell, TableRow } from "@mui/material";
 import {
   FetchNextPageOptions,
@@ -18,7 +17,6 @@ import {
 import type { NumbersViewData } from "@utils/api/endpoints/numbers-view/numbers";
 import type { APIGetResponse, APIPostResponse } from "@utils/api/types";
 import { DbLanguage } from "@utils/api/types";
-import { useSetAtom } from "jotai";
 
 import {
   createTableColumns,
@@ -61,16 +59,13 @@ export default function NumbersTable({
 
   const rowData = React.useMemo(() => createTableRows(data), [data]);
 
-  const setCurrentDbFileAtom = useSetAtom(currentDbFileAtom);
-
   const columns = React.useMemo<ColumnDef<NumbersSegment>[]>(
     () =>
       createTableColumns({
         categories,
         language,
-        setCurrentDbFileAtom,
       }),
-    [categories, language, setCurrentDbFileAtom],
+    [categories, language],
   );
 
   const table = useReactTable({
