+++ frontend/src/features/numbersView/numbersViewTableContent.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { TableComponents } from "react-virtuoso";
-import { SetCurrentDbFileAtom } from "@atoms/types";
 import { Link } from "@components/common/Link";
 import { createURLToSegment } from "@features/textView/utils";
 import {
@@ -47,13 +46,11 @@ export const createTableRows = (
 interface CreateTableColumnProps {
   categories: APIGetResponse<"/numbers-view/categories/">;
   language: DbLanguage;
-  setCurrentDbFileAtom: SetCurrentDbFileAtom;
 }
 
 export const createTableColumns = ({
   categories,
   language,
-  setCurrentDbFileAtom,
 }: CreateTableColumnProps): ColumnDef<NumbersSegment>[] => [
   {
     accessorKey: "segment",
@@ -73,7 +70,6 @@ export const createTableColumns = ({
         <Typography sx={{ fontWeight: 500, lineHeight: 1.25 }}>
           <Link
             href={createURLToSegment({ segmentNumber: segmentnr, language })}
-            onClick={() => setCurrentDbFileAtom(null)}
           >
             {segmentnr}
           </Link>
@@ -123,11 +119,7 @@ export const createTableColumns = ({
                 enterDelay={1200}
               >
                 <Typography sx={{ lineHeight: 1.25 }}>
-                  <Link
-                    href={urlToSegment}
-                    color="text.primary"
-                    onClick={() => setCurrentDbFileAtom(null)}
-                  >
+                  <Link href={urlToSegment} color="text.primary">
                     {segmentnr}
                   </Link>
                 </Typography>
